Extract SyncController from LibraryManager
The cloud-sync responsibility — the upload pipeline (outbox in-flight,
throughput, pause), the connection lifecycle and provider configuration,
the encryption-service cell, membership, and the coven make-Remote /
make-Local primitives — moves out of LibraryManager into a composed
SyncController that holds clones of the handles those paths need and the
five sync-state fields. LibraryManager keeps one `sync: SyncController`
and delegates its public sync API to it, with every public signature
unchanged so bae-bridge and the transfer layer are untouched.

The controller never points back at the manager. The few sync entry
points that also drive the resolver (re-emitting every album after a
cloud-home connect/disconnect) keep their public method on the manager,
which calls the controller for the sync work and does the re-emit itself.
The download/pin queue and foreground transfers stay on the manager: they
run through `TransferService::new(LibraryManager)` and the manager's blob
I/O, so moving only their state would be indirection without ownership.
