+++ bae-core/src/library/manager/sync.rs
@@ -10,11 +10,11 @@ impl LibraryManager {
     /// first read, whereas the `SyncManager` (and its cloud client) is built lazily
     /// once connected and still absent on mobile when the first listing query runs.
     pub fn is_sync_configured(&self) -> bool {
-        self.config_handle.config().cloud_home.provider.is_some()
+        self.sync.is_sync_configured()
     }
 
     pub fn has_cloud_home(&self) -> bool {
-        self.sync_connected()
+        self.sync.has_cloud_home()
     }
 
     /// The coven `BlobRef` for a remote release file's audio blob — its identity
@@ -221,20 +221,13 @@ impl LibraryManager {
     /// entries, not mid-write). Re-emits the outbox snapshot so the UI's
     /// paused indicator and the bottom-panel summary update.
     pub async fn set_sync_paused(&self, paused: bool) {
-        self.sync_paused
-            .store(paused, std::sync::atomic::Ordering::SeqCst);
-        if !paused {
-            // Kick the loop so the queue starts draining immediately on resume
-            // rather than waiting for the next idle tick.
-            self.trigger_sync();
-        }
-        self.emit_outbox_changed().await;
+        self.sync.set_sync_paused(paused).await
     }
 
     /// Current paused state of the upload pipeline. The snapshot builder
     /// reads this so the UI can render its paused indicator.
     pub fn is_sync_paused(&self) -> bool {
-        self.sync_paused.load(std::sync::atomic::Ordering::SeqCst)
+        self.sync.is_sync_paused()
     }
 
     // Sync provider configuration
@@ -244,58 +237,17 @@ impl LibraryManager {
     /// release only becomes remote once the upload observer (which fires from
     /// inside the running loop) confirms the last upload landed.
     pub fn is_sync_ready(&self) -> bool {
-        self.sync_connected() && self.handle.is_syncing()
+        self.sync.is_sync_ready()
     }
 
     pub fn trigger_sync(&self) {
-        self.handle.sync_now();
+        self.sync.trigger_sync()
     }
 
     pub async fn save_s3_config(&self, data: S3ConfigData) -> Result<(), String> {
-        use crate::keys::CloudHomeCredentials;
-        use coven::CloudHome;
-        use coven::S3CloudHome;
-
-        // Probe the bucket with the proposed credentials *before* persisting
-        // anything. A typo or a missing bucket would otherwise leave the UI
-        // showing "Connected" and the user discovering broken sync only via
-        // the reconnect banner after the first failed cycle.
-        let probe_home = S3CloudHome::new(
-            data.bucket.clone(),
-            data.region.clone(),
-            data.endpoint.clone(),
-            data.access_key.clone(),
-            data.secret_key.clone(),
-            data.key_prefix.clone(),
-        )
-        .await
-        .map_err(|e| format!("Failed to build S3 client: {e}"))?;
-        probe_home.probe().await.map_err(|e| format!("{e}"))?;
-
-        let creds = CloudHomeCredentials::S3 {
-            access_key: data.access_key,
-            secret_key: data.secret_key,
-        };
-        self.key_service
-            .set_cloud_home_credentials(&creds)
-            .map_err(|e| format!("Failed to save credentials: {e}"))?;
-
-        self.config_handle
-            .update(move |c| {
-                c.cloud_home.provider = Some(CloudProvider::S3);
-                c.cloud_home.s3_bucket = Some(data.bucket);
-                c.cloud_home.s3_region = Some(data.region);
-                c.cloud_home.s3_endpoint = data.endpoint.filter(|s| !s.is_empty());
-                c.cloud_home.s3_key_prefix = data.key_prefix.filter(|s| !s.is_empty());
-                c.cloud_home.storage = data.storage;
-            })
-            .map_err(|e| format!("Failed to save config: {e}"))?;
-
-        self.ensure_sync_manager_and_start().await?;
+        self.sync.save_s3_config(data).await?;
         // A cloud home now exists, so every release gains its storage actions.
         self.emit_all_albums_updated().await;
-
-        info!("Saved S3 sync configuration");
         Ok(())
     }
 
@@ -305,98 +257,23 @@ impl LibraryManager {
         provider: CloudProvider,
         storage: crate::config::HomeStorage,
     ) -> Result<(), String> {
-        use coven::{sign_in_dropbox, sign_in_google_drive, sign_in_onedrive};
-
-        // Hold the sender alive across the await so cancel.wait_for inside
-        // oauth::authorize never fires (this fn doesn't surface a cancel
-        // signal). When this future is dropped, oauth::authorize's own
-        // AbortOnDrop guard tears the listener task down.
-        let (_cancel_tx, cancel_rx) = tokio::sync::watch::channel(false);
-
-        let library_name = self.config_handle.config().library_name.clone();
-        let clock = self.clock.as_ref();
-        // coven's sign-ins authorize, resolve the cloud folder, and save tokens to
-        // the keyring, returning the folder identifiers; bae persists them here.
-        match provider {
-            CloudProvider::GoogleDrive => {
-                let folder_id =
-                    sign_in_google_drive(&self.key_service, &library_name, cancel_rx, clock)
-                        .await
-                        .map_err(|e| e.to_string())?;
-                self.config_handle
-                    .update(move |c| {
-                        c.cloud_home.provider = Some(CloudProvider::GoogleDrive);
-                        c.cloud_home.google_drive_folder_id = Some(folder_id);
-                        c.cloud_home.storage = storage;
-                    })
-                    .map_err(|e| e.to_string())?;
-            }
-            CloudProvider::Dropbox => {
-                let folder_path =
-                    sign_in_dropbox(&self.key_service, &library_name, cancel_rx, clock)
-                        .await
-                        .map_err(|e| e.to_string())?;
-                self.config_handle
-                    .update(move |c| {
-                        c.cloud_home.provider = Some(CloudProvider::Dropbox);
-                        c.cloud_home.dropbox_folder_path = Some(folder_path);
-                        c.cloud_home.storage = storage;
-                    })
-                    .map_err(|e| e.to_string())?;
-            }
-            CloudProvider::OneDrive => {
-                let (drive_id, folder_id) = sign_in_onedrive(&self.key_service, cancel_rx, clock)
-                    .await
-                    .map_err(|e| e.to_string())?;
-                self.config_handle
-                    .update(move |c| {
-                        c.cloud_home.provider = Some(CloudProvider::OneDrive);
-                        c.cloud_home.onedrive_drive_id = Some(drive_id);
-                        c.cloud_home.onedrive_folder_id = Some(folder_id);
-                        c.cloud_home.storage = storage;
-                    })
-                    .map_err(|e| e.to_string())?;
-            }
-            _ => return Err("This provider does not use OAuth sign-in".to_string()),
-        }
-        self.ensure_sync_manager_and_start().await?;
+        self.sync
+            .sign_in_cloud_provider(provider, storage, &self.clock)
+            .await?;
         // A cloud home now exists, so every release gains its storage actions.
         self.emit_all_albums_updated().await;
         Ok(())
     }
 
     pub async fn use_cloudkit(&self, storage: crate::config::HomeStorage) -> Result<(), String> {
-        self.config_handle
-            .update(move |c| {
-                c.cloud_home.provider = Some(CloudProvider::CloudKit);
-                c.cloud_home.storage = storage;
-            })
-            .map_err(|e| format!("Failed to save CloudKit config: {e}"))?;
-        self.ensure_sync_manager_and_start().await?;
+        self.sync.use_cloudkit(storage).await?;
         // A cloud home now exists, so every release gains its storage actions.
         self.emit_all_albums_updated().await;
-
-        info!("Configured CloudKit cloud provider");
         Ok(())
     }
 
     pub fn disconnect_cloud_provider(&self) -> Result<(), String> {
-        // Stop the sync loop and drop the installed manager; the library becomes
-        // home-less until the next connect.
-        self.handle.disconnect_sync();
-        self.sync_connected
-            .store(false, std::sync::atomic::Ordering::SeqCst);
-        *self.encryption_service.write().unwrap() = None;
-
-        // Connecting fills the whole cloud home; disconnecting clears it as a unit.
-        self.config_handle
-            .update(|c| c.cloud_home = Default::default())
-            .map_err(|e| e.to_string())?;
-
-        // Clearing the cloud-home credentials from the keyring is coven's concern.
-        if let Err(e) = self.key_service.delete_cloud_home_credentials() {
-            tracing::warn!("Failed to delete cloud home credentials: {e}");
-        }
+        self.sync.disconnect_cloud_provider()?;
 
         // The cloud home is gone, so releases lose their storage actions; re-emit
         // every album so cached UI details drop the now-invalid actions. Spawned
@@ -446,77 +323,13 @@ impl LibraryManager {
 
     /// Build, start, and attach a sync manager. Used once at startup for a
     /// returning user with a configured cloud home: an unlocked key for an opaque
-    /// home (`Some`), or no key for a browsable home (`None`). Shares this manager's
-    /// outbox in-flight set and event channel with the sync loop's upload
-    /// observer. Call before [`Self::start`].
+    /// home (`Some`), or no key for a browsable home (`None`). Shares the sync
+    /// controller's outbox in-flight set and event channel with the sync loop's
+    /// upload observer. Call before [`Self::start`].
     pub async fn attach_and_start_sync(
         &self,
         encryption_service: Option<EncryptionService>,
     ) -> Result<(), String> {
-        self.handle.connect_sync(encryption_service.clone()).await?;
-        *self.encryption_service.write().unwrap() = encryption_service;
-        self.sync_connected
-            .store(true, std::sync::atomic::Ordering::SeqCst);
-        Ok(())
-    }
-
-    /// Ensure a SyncManager exists (creating encryption key if needed) and start sync.
-    async fn ensure_sync_manager_and_start(&self) -> Result<(), String> {
-        // If we already have a sync manager, just (re)start its loop.
-        if self.sync_connected() {
-            self.handle.start_sync().await?;
-            return Ok(());
-        }
-
-        // An opaque home mints (or reuses) the library key and seals every object
-        // under it; a browsable home stores in the clear and has no key at all.
-        // Build the encryption service only for an opaque home, so a browsable
-        // home never mints a key it would never use. `get_or_create_encryption_key`
-        // is idempotent, so a retry after a failed sync init reuses the key.
-        let storage = self.config_handle.config().cloud_home.storage;
-        let (enc_service, fingerprint) = if storage.is_opaque() {
-            let enc_key_hex = self
-                .key_service
-                .get_or_create_encryption_key()
-                .map_err(|e| format!("Failed to create encryption key: {e}"))?;
-            let enc = EncryptionService::new(&enc_key_hex)
-                .map_err(|e| format!("Failed to create encryption service: {e}"))?;
-            let fingerprint = enc.fingerprint();
-            (Some(enc), Some(fingerprint))
-        } else {
-            (None, None)
-        };
-
-        // Connect the provider: build the cloud home, start the loop, and install
-        // the manager. A cloud-home build or loop-start failure returns `Err` with
-        // nothing installed, so it surfaces here rather than leaving a dead manager
-        // — and the encryption-key fingerprint below is reached only on success, so
-        // a failed setup stays a clean retry (no fingerprint telling the next
-        // launch's unlock flow "encryption is set up" while sync is still broken).
-        self.handle.connect_sync(enc_service.clone()).await?;
-        *self.encryption_service.write().unwrap() = enc_service;
-        self.sync_connected
-            .store(true, std::sync::atomic::Ordering::SeqCst);
-
-        // Sync started. For an opaque home, persist the encryption-key hint flag so
-        // the next launch's unlock flow knows this library has encryption set up. A
-        // browsable home records nothing — it has no key, so `encryption_key_stored`
-        // stays false and the next launch builds it keyless.
-        if let Some(fingerprint) = fingerprint {
-            if let Err(e) = self
-                .config_handle
-                .record_encryption_key_fingerprint(fingerprint)
-            {
-                self.handle.disconnect_sync();
-                *self.encryption_service.write().unwrap() = None;
-                self.sync_connected
-                    .store(false, std::sync::atomic::Ordering::SeqCst);
-                return Err(format!("Failed to save config: {e}"));
-            }
-        }
-
-        self.trigger_sync();
-
-        Ok(())
+        self.sync.attach_and_start_sync(encryption_service).await
     }
 }
